Separate testclass to add listeners
